move data loading to different module
